formatColumn now a function within qualityControlIndex
